Make exporter `timeout` encompass retries/backoffs, add jitter to backoffs, cleanup code a bit (#4564)
* Initial commit to add timeout as a parm to export, make retries encompass timeout

* Fix lint issues

* Fix a bunch of failing style/lint/spellcheck checks

* Remove timeout param from the export calls.

* Fix flaky windows test ?

* Respond to review comments..

* Delete exponential backoff code that is now unused

* Add changelog and remove some unused imports..

* fix typo and unit test flaking on windows

* Refactor tests, HTTP exporters a bit

* Remove unneeded test reqs

* Remove gRPC retry config

* Tweak backoff calculation

* Lint and precommit

* Empty commit

* Another empty commit

* Calculate backoff in 1 place instead of 2

* Update changelog

* Update changelog

* Make new _common directory in the http exporter for shared code

* precommit

* Respond to comments on PR

* Fix broken test, execute precommit

* Skip some tests on windows

* Explain why test is skipped

* Update exporter/opentelemetry-exporter-otlp-proto-grpc/src/opentelemetry/exporter/otlp/proto/grpc/exporter.py


* Revert change to start respecting timeout passed into metric exporter

---------

+++ exporter/opentelemetry-exporter-otlp-proto-grpc/src/opentelemetry/exporter/otlp/proto/grpc/metric_exporter/__init__.py
@@ -99,7 +99,7 @@ class OTLPMetricExporter(
         credentials: ChannelCredentials | None = None,
         headers: Union[TypingSequence[Tuple[str, str]], dict[str, str], str]
         | None = None,
-        timeout: int | None = None,
+        timeout: float | None = None,
         compression: Compression | None = None,
         preferred_temporality: dict[type, AggregationTemporality]
         | None = None,
@@ -124,7 +124,7 @@ class OTLPMetricExporter(
 
         environ_timeout = environ.get(OTEL_EXPORTER_OTLP_METRICS_TIMEOUT)
         environ_timeout = (
-            int(environ_timeout) if environ_timeout is not None else None
+            float(environ_timeout) if environ_timeout is not None else None
         )
 
         compression = (
@@ -172,7 +172,6 @@ class OTLPMetricExporter(
 
             if split_export_result is MetricExportResult.FAILURE:
                 export_result = MetricExportResult.FAILURE
-
         return export_result
 
     def _split_metrics_data(
